--- Content.Goobstation.Shared/Phones/Components/MobilePhoneComponent.cs
@@ -1,4 +1,0 @@
-namespace Content.Goobstation.Shared.Phones.Components;
-
-[RegisterComponent]
-public sealed partial class MobilePhoneComponent : Component;
